asset-preload => asset-preload-test
